Policies - Implement policy and gates

+++ app/Http/Controllers/Api/PackageController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\PackageRequest;
+use App\Http\Requests\UpdatePackageRequest;
 use App\Http\Resources\V1\PackageResource;
 use App\Models\V1\Package;
+use Illuminate\Support\Facades\Gate;
 
 class PackageController extends Controller
 {
@@ -14,6 +16,8 @@ class PackageController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Package::class);
+
         $packages = Package::with('customer', 'driver', 'address')
             ->latest('id')
             ->paginate(15)
@@ -37,15 +41,21 @@ class PackageController extends Controller
      */
     public function show(Package $package)
     {
+        Gate::authorize('view', $package);
+
         return new PackageResource($package->loadMissing('customer', 'driver', 'address'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(PackageRequest $request, Package $package)
+    public function update(UpdatePackageRequest $request, Package $package)
     {
-        $package->update($request->validated());
+        if (auth()->user()->isDriver()){
+            $package->update(['status' => $request->status]);
+        }else{
+            $package->update($request->validated());
+        }
 
         return new PackageResource($package->loadMissing('customer', 'driver', 'address'));
     }
@@ -55,6 +65,8 @@ class PackageController extends Controller
      */
     public function destroy(Package $package)
     {
+        Gate::authorize('delete', $package);
+
         $package->delete();
 
         return response()->json([
